refactor: rename fake service
